+++ board/board/__init__.py
@@ -24,6 +24,6 @@ def main(global_config, **settings):
     config.add_static_view('static', 'board:static')
     # Configure routes
     config.scan(views)
-    config.include(views.add_routes)
+    config.include(views)
     # Return WSGI app
     return config.make_wsgi_app()