+++ auth/CHANGES.rst
@@ -4,4 +4,4 @@
 - Made tests for public, protected, privileged pages
 - Made data models
 - Made public, protected, privileged pages
-- Made login_ and logout_ views
+- Made login and logout views